Revert "Made bottom nav bar from template"
