--- app/src/main/java/com/example/bna/ui/settings/SettingsViewModel.java
@@ -1,19 +1,0 @@
-package com.example.bna.ui.settings;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class SettingsViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public SettingsViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is settings fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
