--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_chat">
-
-    <fragment
-        android:id="@+id/navigation_chat"
-        android:name="com.example.bna.ui.home.HomeFragment"
-        android:label="@string/title_chat"
-        tools:layout="@layout/fragment_home" />
-
-    <fragment
-        android:id="@+id/navigation_forum"
-        android:name="com.example.bna.ui.dashboard.DashboardFragment"
-        android:label="@string/title_forum"
-        tools:layout="@layout/fragment_dashboard" />
-
-    <fragment
-        android:id="@+id/navigation_search"
-        android:name="com.example.bna.ui.notifications.NotificationsFragment"
-        android:label="@string/title_search"
-        tools:layout="@layout/fragment_notifications" />
-    <fragment
-        android:id="@+id/navigation_settings"
-        android:name="com.example.bna.ui.settings.SettingsFragment"
-        android:label="@string/title_settings"
-        tools:layout="@layout/fragment_settings" />
-</navigation>
