+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,3 @@
 <resources>
     <string name="app_name">JetMentor</string>
-    <string name="title_chat">Chat</string>
-    <string name="title_forum">Forum</string>
-    <string name="title_search">Search</string>
-    <string name="title_settings">Settings</string>
 </resources>
