+++ docs/Bios.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>Bios</title>
+		<link rel="stylesheet" type="text/css" href="main.css">
+		<!-- <link rel="stylesheet" type="text/css" href="/Users/richardboone/Documents/GitHub/jesseboone.github.io/main.css"> -->
+	</head>
+	<body>
+		<nav>
+    		<ul>
+        		<li><a href="./index.html">Home</a></li>
+        		<li><a href="./Goals.html">Goals</a></li>
+        		<li><a href="./Deliverables.html">Deliverables</a></li>
+	        	<li><a href="./Bios.html">Bios</a></li>
+    		</ul>
+		</nav>
+		<div class="container">
+    		<div class="blurb">
+        		<h1>Bios</h1>
+				
+				<h2>Jesse Boone</h2>
+        		<img src="img/IMG_9591.JPG" width="500" height="333">
+        		<p> A sorta grad student here at the University of Alabama.  Plays rugby.</p>
+				
+				<h2>Nick Haeff</h2>
+				<img src="img/Nick-Fall2018-UA.jpeg" width="320" height="400">
+				<p> Senior here at UA.  Interests include video games and going to the gym. </p>
+
+        		<h2>Phil Hartman</h2>
+				<img src="img/HeadshotPhilipHartman.png" width="400" height="600">
+				<p> A senior in Computer Science at The University of Alabama. Interests include homebrewing and going to the gym. </p>
+
+				<h2>Robbie Muma</h2>
+				<img class="image" src="img/Robbie-Muma.jpg" width="320" height="">
+				<p> Senior at University of Alabama. Interests include watching and playing sports.  </p>
+			
+    		</div><!-- /.blurb -->
+		</div><!-- /.container -->
+		<footer>
+    		<ul>
+        		<!-- <li><a href="mailto:hankquinlanhub@gmail.com">email</a></li> -->
+        		<!-- <li><a href="https://github.com/jesseboone">github.com/jesseboone</a></li> -->
+			</ul>
+		</footer>
+	</body>
+</html>
