+++ docs/Deliverables.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>Deliverables</title>
+		<link rel="stylesheet" type="text/css" href="main.css"> 
+		<!-- <link rel="stylesheet" type="text/css" href="/Users/richardboone/Documents/GitHub/jesseboone.github.io/main.css"> -->
+	</head>
+	<body>
+		<nav>
+    		<ul>
+        		<li><a href="./index.html">Home</a></li>
+        		<li><a href="./Goals.html">Goals</a></li>
+        		<li><a href="./Deliverables.html">Deliverables</a></li>
+	        	<li><a href="./Bios.html">Bios</a></li>
+    		</ul>
+		</nav>
+		<div class="container">
+    		<div class="blurb">
+        		<h1>Deliverables</h1>
+        		<h2>Source Code</h2>
+        		<h2>APK</h2>
+        		<h2>Presentation</h2>
+        		<h2>Delivered Documents</h2>
+				<p> </p>
+    		</div><!-- /.blurb -->
+		</div><!-- /.container -->
+		<footer>
+    		<ul>
+        		<!-- <li><a href="mailto:hankquinlanhub@gmail.com">email</a></li> -->
+        		<!-- <li><a href="https://github.com/jesseboone">github.com/jesseboone</a></li> -->
+			</ul>
+		</footer>
+	</body>
+</html>
