+++ docs/Goals.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>Goals</title>
+		<link rel="stylesheet" type="text/css" href="main.css"> 
+		<!-- <link rel="stylesheet" type="text/css" href="/Users/richardboone/Documents/GitHub/jesseboone.github.io/main.css"> -->
+	</head>
+	<body>
+		<nav>
+    		<ul>
+        		<li><a href="./index.html">Home</a></li>
+        		<li><a href="./Goals.html">Goals</a></li>
+        		<li><a href="./Deliverables.html">Deliverables</a></li>
+	        	<li><a href="./Bios.html">Bios</a></li>
+    		</ul>
+		</nav>
+		<div class="container">
+    		<div class="blurb">
+        		<h1>Goals of the project</h1>
+				<p>The goal for this app is to connect prospective employees with company mentors to train them up for the interview process </p>
+    		</div><!-- /.blurb -->
+		</div><!-- /.container -->
+		<footer>
+    		<ul>
+        		<!-- <li><a href="mailto:hankquinlanhub@gmail.com">email</a></li> -->
+        		<!-- <li><a href="https://github.com/jesseboone">github.com/jesseboone</a></li> -->
+			</ul>
+		</footer>
+	</body>
+</html>
