+++ docs/index.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<title>JetMentor</title>
+		<!-- link to main stylesheet -->
+		<link rel="stylesheet" type="text/css" href="main.css"> 
+		<!-- <link rel="stylesheet" type="text/css" href="/Users/richardboone/Documents/GitHub/jesseboone.github.io/main.css"> -->
+	</head>
+	<body>
+		<nav>
+    		<ul>
+        		<li><a href="./Goals.html">Goals</a></li>
+        		<li><a href="./Deliverables.html">Deliverables</a></li>
+	        	<li><a href="./Bios.html">Bios</a></li>
+    		</ul>
+		</nav>
+		<div class="container">
+    		<div class="blurb">
+        		<h1>Welcome to the JetMentor Website!</h1>
+				<p>The goal for this app is to connect prospective employees with company mentors to train them up for the interview process ---- <a href="./Goals.html">Read more about the goals here...</a></p>
+				<p>Click <a href="https://youtu.be/dQw4w9WgXcQ">here</a> to see a demonstration of our app!</p>
+
+			</div><!-- /.blurb -->
+		</div><!-- /.container -->
+		<footer>
+    		<ul>
+        		<!-- <li><a href="mailto:hankquinlanhub@gmail.com">email</a></li> -->
+        		<!-- <li><a href="https://github.com/jesseboone">github.com/jesseboone</a></li> -->
+			</ul>
+		</footer>
+	</body>
+</html>
